# apps/api/app/Http/Requests/CreateVehicleTypeRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;
use Illuminate\Validation\Rule;

class CreateVehicleTypeRequest extends FormRequest
{
    public function rules(): array
    {
        return [
            'code' => ['required', 'string', 'max:64', 'regex:/^[A-Z0-9_-]+$/', Rule::unique('vehicle_types')->where('company_id', $this->route('company')->id)],
            'name' => ['required', 'string', 'max:255'],
        ];
    }
}