# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Rajawali Platform API
  version: 0.15.0
  description: Contract-first API shared by ERP Web, Operations Web, and mobile. Every API response returns X-Request-ID; authenticated business mutations may use Idempotency-Key for safe retry.
...
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
paths:
...
  /companies/{companyId}/locations/{locationId}/fleet/vehicle-types:
    parameters:
      - { $ref: "#/components/parameters/CompanyId" }
      - { $ref: "#/components/parameters/LocationId" }
    get:
      operationId: listVehicleTypes
      summary: List active vehicle types for a location-scoped Fleet workspace.
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Active company vehicle types.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/VehicleType" } }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createVehicleType
      summary: Create a company vehicle type.
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateVehicleTypeRequest" }
      responses:
        "201":
          description: Vehicle type created.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/VehicleType" } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /companies/{companyId}/locations/{locationId}/fleet/vehicles:
    parameters:
      - { $ref: "#/components/parameters/CompanyId" }
      - { $ref: "#/components/parameters/LocationId" }
    get:
      operationId: listVehicles
      summary: List vehicles in one authorized operating location.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/VehicleStatus" } }
      responses:
        "200":
          description: Paginated vehicles.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VehiclePage" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createVehicle
      summary: Register a vehicle with an initial available status history.
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateVehicleRequest" }
      responses:
        "201":
          description: Vehicle registered.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/Vehicle" } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
  /companies/{companyId}/locations/{locationId}/fleet/vehicles/{vehicleId}/status:
    parameters:
      - { $ref: "#/components/parameters/CompanyId" }
      - { $ref: "#/components/parameters/LocationId" }
      - { $ref: "#/components/parameters/VehicleId" }
      - { $ref: "#/components/parameters/IdempotencyKey" }
    post:
      operationId: changeVehicleStatus
      summary: Apply a guarded vehicle status transition and append its history.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangeVehicleStatusRequest" }
      responses:
        "200":
          description: Vehicle status changed.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: "#/components/schemas/Vehicle" } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: rajawali_session
    bearerAuth:
      type: http
      scheme: bearer
    metricsAuth:
      type: http
      scheme: bearer
      description: Dedicated OBSERVABILITY_METRICS_TOKEN; employee tokens are not accepted.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: A client-generated key for safely retrying an authenticated business mutation. Reuse with another route or payload returns 409.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{15,127}$"
...
    CompanyId:
      name: companyId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
    LocationId:
      name: locationId
      in: path
      required: true
      schema: { type: string, pattern: "^[0-9A-HJKMNP-TV-Z]{26}$" }
    VehicleId:
      name: vehicleId
      in: path
      required: true
      schema: { type: string, pattern: "^[0-9A-HJKMNP-TV-Z]{26}$" }
...
  responses:
...
    Forbidden:
      description: Authenticated identity is inactive or lacks scoped permission.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationFailed:
      description: Request validation or lifecycle rule failed.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The requested lifecycle transition conflicts with current state.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  headers:
    XRequestId:
      description: Correlation identifier for logs, audit records, and support diagnostics.
      required: true
      schema:
        type: string
...
  schemas:
...
    VehicleStatus:
      type: string
      enum: [available, in_use, maintenance, blocked, inactive]
...
    VehicleType:
      type: object
      required: [id, code, name]
      properties: { id: { type: string }, code: { type: string }, name: { type: string } }
    CreateVehicleTypeRequest:
      type: object
      required: [code, name]
      properties: { code: { type: string, pattern: "^[A-Z0-9_-]+$" }, name: { type: string } }
    Vehicle:
      type: object
      required: [id, code, plate_number, brand, model, current_odometer, operational_status]
      properties:
        id: { type: string }
        code: { type: string }
        plate_number: { type: string }
        brand: { type: string }
        model: { type: string }
        model_year: { type: [integer, "null"] }
        ownership_type: { type: string, enum: [owned, leased, vendor] }
        provider_name: { type: [string, "null"] }
        current_odometer: { type: integer, minimum: 0 }
        operational_status: { $ref: "#/components/schemas/VehicleStatus" }
        status_reason: { type: [string, "null"] }
        type: { $ref: "#/components/schemas/VehicleType" }
    CreateVehicleRequest:
      type: object
      required: [vehicle_type_id, code, plate_number, brand, model, ownership_type, current_odometer]
      properties:
        vehicle_type_id: { type: string }
        code: { type: string }
        plate_number: { type: string }
        brand: { type: string }
        model: { type: string }
        model_year: { type: [integer, "null"] }
        ownership_type: { type: string, enum: [owned, leased, vendor] }
        provider_name: { type: [string, "null"] }
        current_odometer: { type: integer, minimum: 0 }
    ChangeVehicleStatusRequest:
      type: object
      required: [status, reason]
      properties:
        status: { $ref: "#/components/schemas/VehicleStatus" }
        reason: { type: string, minLength: 5 }
    VehiclePage:
      type: object
      required: [data, current_page, last_page, total]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/Vehicle" } }
        current_page: { type: integer }
        last_page: { type: integer }
        total: { type: integer }
...
    Error:
      type: object
      required: [message, code, request_id]
      properties:
        message:
          type: string
        code:
          type: string
        request_id:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string